refactor: mv `generateTypechain`
